Sooooo
I think I get it. The way I have it set up, perlin noise is shown as a
bunch of lines that move in order to have a wave-like thing at the top
(or side, depending on how it's set up). The reason that it can appear
as static is not because the lines are moving, but because they're
changing colour. When I rotate them, it shows up as a spiral-like thing
that has lines moving very little back and forth. If you comment out the
red/blue/purple loop, it's easier to see.